requirements.txt
# =============================================
# Requirements — SaaS Churn Prediction Engine
# =============================================
# Core
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# ML
scikit-learn>=1.3.0,<2.0.0
xgboost>=1.7.0
lightgbm>=4.0.0
optuna>=3.0.0
aif360[Reductions]>=0.5.0
imbalanced-learn>=0.11.0
mlflow>=2.0.0
shap>=0.42.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Serialization
joblib>=1.3.0,<2.0.0

# Config
pyyaml>=6.0,<7.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Dev / Testing
pytest>=7.0.0,<9.0.0
pytest-cov>=4.0.0
